Okfn image plugin refact for storing index in the ShapeSelector
Now images are organized in a two-level map structure, with keys: image source and index. This is due to the enhancement to handle multiple images with the same source in a page.
Also, annotorious side annotations store a reference to their imageplugin, the name of the reference property is 'handler'